% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiUnion.R
\docType{package}
\name{multiUnion-package}
\title{multiUnion-package}
\description{
Set operations that take multiple arguments.
}
\author{
\strong{Maintainer}: Dale \email{dale@piratepress.org} [copyright holder]

}
